+++ src/validators/collections/length/index.ts
@@ -1,4 +1,5 @@
 export { IsMinLength } from "./is-min-length";
+export { IsMaxLength } from "./is-max-length";
 
 export { IsEmpty } from "./is-empty";
 export { IsNotEmpty } from "./is-not-empty";